Added processing of repetition of mortgages, visual adjsustments. BETA 0.1.1! .

+++ BudgetProWeb/src/main/webapp/WEB-INF/pages/categories.jsp
@@ -40,28 +40,23 @@
                 <div class="sidebar-wrapper">
                     <ul>
                         <li>
-                            <a href="${pageContext.request.contextPath}/dashboard" data-toggle="tooltip" data-placement="right" title="Startpagina" data-original-title="Startpagina">
+                            <a href="${pageContext.request.contextPath}/dashboard" data-toggle="tooltip" data-placement="right" title="" data-original-title="Startpagina">
                                 <i class="fa fa-home"></i>
                             </a>
                         </li>
                         <li>
                             <a href="${pageContext.request.contextPath}/transaction/list" data-toggle="tooltip" data-placement="right" title="Transacties" data-original-title="Transacties">
-                                <i class="fa fa-map-marker"></i>
+                                <i class="fa fa-euro"></i>
                             </a>
                         </li>
                         <li>
                             <a href="${pageContext.request.contextPath}/category/list" data-toggle="tooltip" data-placement="right" title="Categorieën" data-original-title="Categorieën">
-                                <i class="fa fa-group"></i>
+                                <i class="fa fa-list"></i>
                             </a>
                         </li>
                         <li>
-                            <a href="${pageContext.request.contextPath}" data-toggle="tooltip" data-placement="right" title="" data-original-title="Startpagina">
-                                <i class="fa fa-pencil"></i>
-                            </a>
-                        </li>
-                        <li>
-                            <a href="${pageContext.request.contextPath}" data-toggle="tooltip" data-placement="right" title="" data-original-title="Startpagina">
-                                <i class="fa fa-flash"></i>
+                            <a href="${pageContext.request.contextPath}/mortgage/list" data-toggle="tooltip" data-placement="right" title="Hypotheken" data-original-title="Hypotheken">
+                                <i class="fa fa-money"></i>
                             </a>
                         </li>
                     </ul>
@@ -75,7 +70,7 @@
                                 <i class="fa fa-group"></i> Categorieën
                             </div>
                             <div class="widget-content">
-                                <div class="table-responsive">
+                                <div class="table-responsive" style="height: 395px; overflow: auto;">
                                     <table class="table">
                                         <thead>
                                             <tr>
@@ -90,7 +85,7 @@
                                                     <td>${category.name}</td>
                                                     <td>
                                                         <script>
-                                                            if(${category.incoming}){
+                                                            if (${category.incoming}) {
                                                                 document.write("Ingaand");
                                                             } else {
                                                                 document.write("Uitgaand");
@@ -99,7 +94,7 @@
                                                     </td>
                                                     <td>
                                                         <a href="${pageContext.request.contextPath}/category/edit/${category.id}" class="btn btn-iconed btn-primary btn-xs"><i class="fa fa-pencil"></i>Bewerken</a>
-                                                        <a href="${pageContext.request.contextPath}/category/delete/${category.id}" class="btn btn-danger btn-xs"><i class="fa fa-times"></i>Verwijderen</a>
+                                                        <a onclick="return confirm('Weet je zeker dat je deze categorie wil verwijderen?')" href="${pageContext.request.contextPath}/category/delete/${category.id}" class="btn btn-danger btn-xs"><i class="fa fa-times"></i>Verwijderen</a>
                                                     </td>
                                                 </tr>
                                             </c:forEach>
@@ -122,7 +117,7 @@
                                     </div>
                                     <div class="form-group">
                                         <label control-label>Naam</label>
-                                        <form:input path="name" type="text" placeholder="${category.name}" class="form-control" />
+                                        <form:input path="name" type="text" placeholder="${category.name}" class="form-control" required="true"/>
                                     </div>
                                     <div class="form-group">
                                         <label control-label>Ingaand/Uitgaand</label>
@@ -133,17 +128,11 @@
                                             <form:radiobutton path="incoming" name="incoming" value="false" label=" Uitgaand" />
                                         </div>
                                     </div>
-                                    <div class="row">
-                                        <div class="col-md-4">
-                                            <div class="form-group">
-                                                <form:button type="submit" class="btn btn-primary">Opslaan</form:button>
-                                                </div>
-                                            </div>
-                                            <div class="col-md-4">
-                                                <div class="form-group">
-                                                    <a href="${pageContext.request.contextPath}/category/list" class="btn  btn-warning">Annuleren</a>
-                                            </div>
+                                    <div class="form-group">
+                                        <form:button type="submit" class="btn btn-primary"><i class="fa fa-save"></i>Opslaan</form:button>
                                         </div>
+                                        <div class="form-group">
+                                            <a href="${pageContext.request.contextPath}/category/list" class="btn  btn-danger"><i class="fa fa-times"></i>Annuleren</a>
                                     </div>
                                 </form:form>
                             </div>
